Add files via upload
The idea is to think bit by bit. When a number has ith bit set, whatever we choose now there will be 2^n - 1 subsets (minus 1 for the empty set) of this particular bit will give contribution to the sum. So the formula comes out- if ith bit is set hence, for this particular bit (2^i)*2^(r-l).  proof: 
    (1 + x)^n = x*nC1 + x^2 * nC2 + x^3 * nC3 and so on.
-> putting x = -1,
 (1 -1)^n  = -nC1 + nC2 -nC3 +nC4 - nC5 and so on.
Now, LHS is zero. From the RHS, all odd combinations move to the LHS, then it will be like this:
 nC1 + nC3 + nC5 and so on = nC2 + nC4 + nC6 and so on = 2^n - 1 (excluding the empty set nC0)